refactor(anbima-data-funds): Split into one class per module
  - Extract AnbimaDataFundsAvailable → anbima_data_funds_available.py
  - Extract AnbimaDataFundsAbout → anbima_data_funds_about.py
  - Extract AnbimaDataFundsHistoric → anbima_data_funds_historic.py
  - Split tests into per-class files → test_anbima_data_funds_*.py
  - Split examples into per-class files → examples/anbima_data_funds_*.py
  - Delete original multi-class files → anbima_data_funds.py
