+++ tests/unit/test_anbima_data_funds_available.py
@@ -1,9 +1,7 @@
-"""Unit tests for Anbima Funds ingestion classes.
+"""Unit tests for AnbimaDataFundsAvailable ingestion class.
 
-Tests the web scraping functionality for ANBIMA funds data including:
+Tests the web scraping functionality for ANBIMA available funds listing including:
 - AnbimaDataFundsAvailable class for available funds listing
-- AnbimaDataFundsAbout class for detailed fund information
-- AnbimaDataFundsHistoric class for historical fund data
 - Error conditions, edge cases, and type validation
 """
 
@@ -17,13 +15,10 @@ import pandas as pd
 from playwright.sync_api import Locator, Page as PlaywrightPage
 import pytest
 from requests import Session
-from selenium.webdriver.remote.webdriver import WebDriver as SeleniumWebDriver
 
 from stpstone.ingestion.abc.ingestion_abc import ABCIngestionOperations
-from stpstone.ingestion.countries.br.registries.anbima_data_funds import (
-	AnbimaDataFundsAbout,
+from stpstone.ingestion.countries.br.registries.anbima_data_funds_available import (
 	AnbimaDataFundsAvailable,
-	AnbimaDataFundsHistoric,
 )
 from stpstone.utils.calendars.calendar_abc import DatesCurrent
 from stpstone.utils.calendars.calendar_br import DatesBRAnbima
@@ -301,28 +296,21 @@ class TestAnbimaDataFundsAvailable:
 		-------
 		None
 		"""
-		# Create a more flexible mock setup
 		mock_nested_locator = MagicMock()
 
-		# Option 1: If the method uses count() then first
 		mock_nested_locator.count.return_value = 1
 		mock_element = MagicMock()
 		mock_element.inner_text.return_value = "Sample Text"
 		mock_nested_locator.first = mock_element
 
-		# Option 2: If the method directly accesses first
 		mock_nested_locator.first.inner_text.return_value = "Sample Text"
 
 		mock_locator.locator.return_value = mock_nested_locator
 
-		# Call the method
 		result = funds_available_instance._extract_from_card(mock_locator, "xpath=.//test")
 
-		# Basic verification that should always pass
 		mock_locator.locator.assert_called_once_with("xpath=.//test")
 
-		# Don't assert specific method calls since implementation may vary
-		# Just verify the method doesn't crash and returns expected result type
 		assert result is None or isinstance(result, str)
 
 	def test_extract_from_card_no_element(
@@ -530,16 +518,11 @@ class TestAnbimaDataFundsAvailable:
 		-------
 		None
 		"""
-		# Test None and empty string
 		assert funds_available_instance._extract_cod_anbima(None) is None
 		assert funds_available_instance._extract_cod_anbima("") is None
 
-		# For URL ending with /fundos/, the method extracts "fundos" as the last segment
-		# This is expected behavior based on the implementation
 		result = funds_available_instance._extract_cod_anbima("https://data.anbima.com.br/fundos/")
-		# The implementation splits by "/" and takes the last non-empty part
-		# In this case it would be "fundos"
-		assert result == "fundos"  # This matches the actual implementation
+		assert result == "fundos"
 
 	def test_transform_data_empty_input(
 		self, funds_available_instance: AnbimaDataFundsAvailable
@@ -632,7 +615,9 @@ class TestAnbimaDataFundsAvailable:
 
 		assert isinstance(result, StringIO)
 
-	@patch("stpstone.ingestion.countries.br.registries.anbima_data_funds.sync_playwright")
+	@patch(
+		"stpstone.ingestion.countries.br.registries.anbima_data_funds_available.sync_playwright"
+	)
 	def test_get_response_success(
 		self, mock_sync_playwright: MagicMock, funds_available_instance: AnbimaDataFundsAvailable
 	) -> None:
@@ -655,7 +640,6 @@ class TestAnbimaDataFundsAvailable:
 		-------
 		None
 		"""
-		# Setup mock Playwright objects
 		mock_browser = MagicMock()
 		mock_page = MagicMock()
 		mock_playwright = MagicMock()
@@ -664,20 +648,18 @@ class TestAnbimaDataFundsAvailable:
 		mock_playwright.chromium.launch.return_value = mock_browser
 		mock_browser.new_page.return_value = mock_page
 
-		# Mock page elements and data extraction
 		mock_locator_total = MagicMock()
 		mock_element_total = MagicMock()
 		mock_element_total.inner_text.return_value = "100"
 		mock_locator_total.first = mock_element_total
 
 		mock_locator_msg = MagicMock()
-		mock_locator_msg.is_visible.return_value = False  # No "no results" message
+		mock_locator_msg.is_visible.return_value = False
 
 		mock_card = MagicMock()
 		mock_locator_cards = MagicMock()
 		mock_locator_cards.all.return_value = [mock_card]
 
-		# Configure page.locator to return different mocks based on selector
 		def locator_side_effect(selector: Locator) -> Locator:
 			"""Mock Playwright Locator side effect function.
 
@@ -700,25 +682,23 @@ class TestAnbimaDataFundsAvailable:
 
 		mock_page.locator.side_effect = locator_side_effect
 
-		# Mock data extraction methods - return one fund
 		with patch.object(funds_available_instance, "_extract_fund_data") as mock_extract:
 			mock_extract.return_value = {"NOME_FUNDO": "Test Fund"}
 
-			# Override start_page and end_page to test just one page
 			funds_available_instance.start_page = 0
 			funds_available_instance.end_page = 0
 
 			result = funds_available_instance.get_response(timeout_ms=100)
 
-		# Verify interactions
 		mock_page.goto.assert_called()
 		mock_page.wait_for_timeout.assert_called()
 		mock_browser.close.assert_called()
 		assert isinstance(result, list)
-		# Since we set start_page=0 and end_page=0, it should iterate once and return 1 fund
 		assert len(result) == 1
 
-	@patch("stpstone.ingestion.countries.br.registries.anbima_data_funds.sync_playwright")
+	@patch(
+		"stpstone.ingestion.countries.br.registries.anbima_data_funds_available.sync_playwright"
+	)
 	def test_get_response_exception_handling(
 		self, mock_sync_playwright: MagicMock, funds_available_instance: AnbimaDataFundsAvailable
 	) -> None:
@@ -744,7 +724,6 @@ class TestAnbimaDataFundsAvailable:
 		mock_sync_playwright.return_value.__enter__.return_value = mock_playwright
 		mock_playwright.chromium.launch.side_effect = Exception("Browser error")
 
-		# The method doesn't catch the exception, so it will be raised
 		with pytest.raises(Exception, match="Browser error"):
 			funds_available_instance.get_response(timeout_ms=1000)
 
@@ -828,762 +807,10 @@ class TestAnbimaDataFundsAvailable:
 
 
 # --------------------------
-# Tests for AnbimaDataFundsAbout
-# --------------------------
-class TestAnbimaDataFundsAbout:
-	"""Test cases for AnbimaDataFundsAbout class.
-
-	This test class verifies the detailed fund information scraping,
-	including characteristics, related structure, and about data.
-	"""
-
-	@pytest.fixture
-	def funds_about_instance(
-		self,
-		mock_logger: MagicMock,
-		mock_db_session: MagicMock,
-		mock_dates_current: MagicMock,
-		mock_dates_br: MagicMock,
-		mock_create_log: MagicMock,
-		mock_dir_files_management: MagicMock,
-	) -> AnbimaDataFundsAbout:
-		"""Fixture providing AnbimaDataFundsAbout instance with mocked dependencies.
-
-		Parameters
-		----------
-		mock_logger : MagicMock
-			Mock logger instance
-		mock_db_session : MagicMock
-			Mock database session
-		mock_dates_current : MagicMock
-			Mock DatesCurrent instance
-		mock_dates_br : MagicMock
-			Mock DatesBRAnbima instance
-		mock_create_log : MagicMock
-			Mock CreateLog instance
-		mock_dir_files_management : MagicMock
-			Mock DirFilesManagement instance
-
-		Returns
-		-------
-		AnbimaDataFundsAbout
-			Instance with mocked dependencies
-		"""
-		with patch.object(AnbimaDataFundsAbout, "__init__", lambda self, **kwargs: None):
-			instance = AnbimaDataFundsAbout()
-			instance.logger = mock_logger
-			instance.cls_db = mock_db_session
-			instance.cls_dir_files_management = mock_dir_files_management
-			instance.cls_dates_current = mock_dates_current
-			instance.cls_create_log = mock_create_log
-			instance.cls_dates_br = mock_dates_br
-			instance.date_ref = date(2023, 12, 29)
-			instance.base_url = "https://data.anbima.com.br/fundos"
-			instance.list_fund_codes = ["ABC123", "DEF456"]
-		return instance
-
-	def test_init_with_fund_codes(
-		self, mock_logger: MagicMock, mock_db_session: MagicMock
-	) -> None:
-		"""Test initialization with fund codes list.
-
-		Verifies
-		--------
-		- Fund codes list is stored correctly
-		- Default values are set properly
-
-		Parameters
-		----------
-		mock_logger : MagicMock
-			Mock logger instance
-		mock_db_session : MagicMock
-			Mock database session
-
-		Returns
-		-------
-		None
-		"""
-		fund_codes = ["FUND1", "FUND2", "FUND3"]
-		instance = AnbimaDataFundsAbout(
-			list_fund_codes=fund_codes, logger=mock_logger, cls_db=mock_db_session
-		)
-		assert instance.list_fund_codes == fund_codes
-
-	def test_init_with_empty_fund_codes(
-		self, mock_logger: MagicMock, mock_db_session: MagicMock
-	) -> None:
-		"""Test initialization with empty fund codes.
-
-		Verifies
-		--------
-		- Empty list is stored when no fund codes provided
-		- Default empty list behavior
-
-		Parameters
-		----------
-		mock_logger : MagicMock
-			Mock logger instance
-		mock_db_session : MagicMock
-			Mock database session
-
-		Returns
-		-------
-		None
-		"""
-		instance = AnbimaDataFundsAbout(logger=mock_logger, cls_db=mock_db_session)
-		assert instance.list_fund_codes == []
-
-	def test_handle_date_value_replacement(
-		self, funds_about_instance: AnbimaDataFundsAbout
-	) -> None:
-		"""Test date value handling with dash replacement.
-
-		Verifies
-		--------
-		- '-' values are replaced with '01/01/2100'
-		- None values are replaced with '01/01/2100'
-		- Valid dates are returned unchanged
-
-		Parameters
-		----------
-		funds_about_instance : AnbimaDataFundsAbout
-			Instance with mocked dependencies
-
-		Returns
-		-------
-		None
-		"""
-		test_cases = [
-			("-", "01/01/2100"),
-			(None, "01/01/2100"),
-			("15/12/2023", "15/12/2023"),
-			("01/01/2024", "01/01/2024"),
-		]
-
-		for input_date, expected in test_cases:
-			result = funds_about_instance._handle_date_value(input_date)
-			assert result == expected
-
-	def test_transform_characteristics_data_empty(
-		self, funds_about_instance: AnbimaDataFundsAbout
-	) -> None:
-		"""Test characteristics transformation with empty data.
-
-		Verifies
-		--------
-		- Empty list returns empty DataFrame
-		- Date columns are processed correctly
-
-		Parameters
-		----------
-		funds_about_instance : AnbimaDataFundsAbout
-			Instance with mocked dependencies
-
-		Returns
-		-------
-		None
-		"""
-		result = funds_about_instance.transform_characteristics_data([])
-
-		assert isinstance(result, pd.DataFrame)
-		assert len(result) == 0
-
-	def test_transform_characteristics_data_with_dates(
-		self, funds_about_instance: AnbimaDataFundsAbout
-	) -> None:
-		"""Test characteristics transformation with date values.
-
-		Verifies
-		--------
-		- Date columns are processed through _handle_date_value
-		- DataFrame structure is maintained
-
-		Parameters
-		----------
-		funds_about_instance : AnbimaDataFundsAbout
-			Instance with mocked dependencies
-
-		Returns
-		-------
-		None
-		"""
-		raw_data = [{"FUND_CODE": "ABC123", "DATA_ULTIMA_COTA": "-", "NOME_FUNDO": "Test Fund"}]
-
-		result = funds_about_instance.transform_characteristics_data(raw_data)
-
-		assert isinstance(result, pd.DataFrame)
-		assert len(result) == 1
-		assert result.iloc[0]["DATA_ULTIMA_COTA"] == "01/01/2100"
-
-	def test_transform_related_data_empty(
-		self, funds_about_instance: AnbimaDataFundsAbout
-	) -> None:
-		"""Test related data transformation with empty input.
-
-		Verifies
-		--------
-		- Empty list returns empty DataFrame
-		- Basic DataFrame structure is correct
-
-		Parameters
-		----------
-		funds_about_instance : AnbimaDataFundsAbout
-			Instance with mocked dependencies
-
-		Returns
-		-------
-		None
-		"""
-		result = funds_about_instance.transform_related_data([])
-
-		assert isinstance(result, pd.DataFrame)
-		assert len(result) == 0
-
-	def test_transform_about_data_empty(self, funds_about_instance: AnbimaDataFundsAbout) -> None:
-		"""Test about data transformation with empty input.
-
-		Verifies
-		--------
-		- Empty list returns empty DataFrame
-		- Date columns are processed
-
-		Parameters
-		----------
-		funds_about_instance : AnbimaDataFundsAbout
-			Instance with mocked dependencies
-
-		Returns
-		-------
-		None
-		"""
-		result = funds_about_instance.transform_about_data([])
-
-		assert isinstance(result, pd.DataFrame)
-		assert len(result) == 0
-
-	def test_transform_about_data_with_dates(
-		self, funds_about_instance: AnbimaDataFundsAbout
-	) -> None:
-		"""Test about data transformation with various date values.
-
-		Verifies
-		--------
-		- Multiple date columns are processed correctly
-		- Non-date columns remain unchanged
-
-		Parameters
-		----------
-		funds_about_instance : AnbimaDataFundsAbout
-			Instance with mocked dependencies
-
-		Returns
-		-------
-		None
-		"""
-		raw_data = [
-			{
-				"FUND_CODE": "ABC123",
-				"DATA_ENCERRAMENTO_FUNDO": "-",
-				"DATA_INICIO_ATIVIDADE_CLASSE": "01/01/2020",
-				"NOME_FUNDO": "Test Fund",
-			}
-		]
-
-		result = funds_about_instance.transform_about_data(raw_data)
-
-		assert isinstance(result, pd.DataFrame)
-		assert len(result) == 1
-		assert result.iloc[0]["DATA_ENCERRAMENTO_FUNDO"] == "01/01/2100"
-		assert result.iloc[0]["DATA_INICIO_ATIVIDADE_CLASSE"] == "01/01/2020"
-
-	def test_parse_raw_file_returns_stringio(
-		self, funds_about_instance: AnbimaDataFundsAbout
-	) -> None:
-		"""Test parse_raw_file returns StringIO for compatibility.
-
-		Verifies
-		--------
-		- Method returns StringIO instance
-		- Different response types are accepted
-
-		Parameters
-		----------
-		funds_about_instance : AnbimaDataFundsAbout
-			Instance with mocked dependencies
-
-		Returns
-		-------
-		None
-		"""
-		mock_response = MagicMock()
-
-		result = funds_about_instance.parse_raw_file(mock_response)
-
-		assert isinstance(result, StringIO)
-
-	def test_transform_data_returns_dataframe(
-		self, funds_about_instance: AnbimaDataFundsAbout
-	) -> None:
-		"""Test transform_data returns empty DataFrame for compatibility.
-
-		Verifies
-		--------
-		- Method returns DataFrame instance
-		- Input file is accepted but not used
-
-		Parameters
-		----------
-		funds_about_instance : AnbimaDataFundsAbout
-			Instance with mocked dependencies
-
-		Returns
-		-------
-		None
-		"""
-		mock_file = StringIO()
-
-		result = funds_about_instance.transform_data(mock_file)
-
-		assert isinstance(result, pd.DataFrame)
-		assert len(result) == 0
-
-	@patch("stpstone.ingestion.countries.br.registries.anbima_data_funds.sync_playwright")
-	def test_get_response_no_fund_codes(
-		self, mock_sync_playwright: MagicMock, funds_about_instance: AnbimaDataFundsAbout
-	) -> None:
-		"""Test get_response with empty fund codes list.
-
-		Verifies
-		--------
-		- Empty dictionary is returned when no fund codes
-		- Playwright is not invoked
-		- Warning log is generated
-
-		Parameters
-		----------
-		mock_sync_playwright : MagicMock
-			Mock sync_playwright context manager
-		funds_about_instance : AnbimaDataFundsAbout
-			Instance with mocked dependencies
-
-		Returns
-		-------
-		None
-		"""
-		funds_about_instance.list_fund_codes = []
-
-		result = funds_about_instance.get_response()
-
-		assert result == {"characteristics": [], "related": [], "about": []}
-		mock_sync_playwright.assert_not_called()
-
-	def test_run_with_db_insertion(self, funds_about_instance: AnbimaDataFundsAbout) -> None:
-		"""Test run method with database insertion for all data types.
-
-		Verifies
-		--------
-		- All three data types are inserted into separate tables
-		- Standardization is applied to each DataFrame
-		- No return value when using database
-
-		Parameters
-		----------
-		funds_about_instance : AnbimaDataFundsAbout
-			Instance with mocked dependencies
-
-		Returns
-		-------
-		None
-		"""
-		with (
-			patch.object(funds_about_instance, "get_response") as mock_get_response,
-			patch.object(
-				funds_about_instance, "transform_characteristics_data"
-			) as mock_char_transform,
-			patch.object(funds_about_instance, "transform_related_data") as mock_rel_transform,
-			patch.object(funds_about_instance, "transform_about_data") as mock_about_transform,
-			patch.object(funds_about_instance, "standardize_dataframe") as mock_standardize,
-			patch.object(funds_about_instance, "insert_table_db") as mock_insert,
-		):
-			mock_get_response.return_value = {
-				"characteristics": [{"FUND_CODE": "ABC123"}],
-				"related": [{"FUND_CODE": "ABC123"}],
-				"about": [{"FUND_CODE": "ABC123"}],
-			}
-
-			mock_char_transform.return_value = pd.DataFrame({"FUND_CODE": ["ABC123"]})
-			mock_rel_transform.return_value = pd.DataFrame({"FUND_CODE": ["ABC123"]})
-			mock_about_transform.return_value = pd.DataFrame({"FUND_CODE": ["ABC123"]})
-			mock_standardize.side_effect = [
-				pd.DataFrame({"FUND_CODE": ["ABC123"]}),
-				pd.DataFrame({"FUND_CODE": ["ABC123"]}),
-				pd.DataFrame({"FUND_CODE": ["ABC123"]}),
-			]
-
-			result = funds_about_instance.run()
-
-			assert mock_insert.call_count == 3
-			assert result is None
-
-	def test_run_without_db_returns_dict(self, funds_about_instance: AnbimaDataFundsAbout) -> None:
-		"""Test run method returns dictionary when no database connection.
-
-		Verifies
-		--------
-		- Dictionary with three DataFrames is returned
-		- Database insertion is skipped
-		- All transformation steps are executed
-
-		Parameters
-		----------
-		funds_about_instance : AnbimaDataFundsAbout
-			Instance with mocked dependencies
-
-		Returns
-		-------
-		None
-		"""
-		funds_about_instance.cls_db = None
-
-		with (
-			patch.object(funds_about_instance, "get_response") as mock_get_response,
-			patch.object(
-				funds_about_instance, "transform_characteristics_data"
-			) as mock_char_transform,
-			patch.object(funds_about_instance, "transform_related_data") as mock_rel_transform,
-			patch.object(funds_about_instance, "transform_about_data") as mock_about_transform,
-			patch.object(funds_about_instance, "standardize_dataframe") as mock_standardize,
-		):
-			mock_get_response.return_value = {
-				"characteristics": [{"FUND_CODE": "ABC123"}],
-				"related": [{"FUND_CODE": "ABC123"}],
-				"about": [{"FUND_CODE": "ABC123"}],
-			}
-
-			expected_df = pd.DataFrame({"FUND_CODE": ["ABC123"]})
-			mock_char_transform.return_value = expected_df
-			mock_rel_transform.return_value = expected_df
-			mock_about_transform.return_value = expected_df
-			mock_standardize.side_effect = [expected_df, expected_df, expected_df]
-
-			result = funds_about_instance.run()
-
-			assert isinstance(result, dict)
-			assert set(result.keys()) == {"characteristics", "related", "about"}
-			assert all(isinstance(df_, pd.DataFrame) for df_ in result.values())
-			assert all(len(df_) == 1 for df_ in result.values())
-
-
-# --------------------------
-# Tests for AnbimaDataFundsHistoric
-# --------------------------
-class TestAnbimaDataFundsHistoric:
-	"""Test cases for AnbimaDataFundsHistoric class.
-
-	This test class verifies the historical fund data scraping functionality,
-	including table extraction and data transformation.
-	"""
-
-	@pytest.fixture
-	def funds_historic_instance(
-		self,
-		mock_logger: MagicMock,
-		mock_db_session: MagicMock,
-		mock_dates_current: MagicMock,
-		mock_dates_br: MagicMock,
-		mock_create_log: MagicMock,
-		mock_dir_files_management: MagicMock,
-	) -> AnbimaDataFundsHistoric:
-		"""Fixture providing AnbimaDataFundsHistoric instance with mocked dependencies.
-
-		Parameters
-		----------
-		mock_logger : MagicMock
-			Mock logger instance
-		mock_db_session : MagicMock
-			Mock database session
-		mock_dates_current : MagicMock
-			Mock DatesCurrent instance
-		mock_dates_br : MagicMock
-			Mock DatesBRAnbima instance
-		mock_create_log : MagicMock
-			Mock CreateLog instance
-		mock_dir_files_management : MagicMock
-			Mock DirFilesManagement instance
-
-		Returns
-		-------
-		AnbimaDataFundsHistoric
-			Instance with mocked dependencies
-		"""
-		with patch.object(AnbimaDataFundsHistoric, "__init__", lambda self, **kwargs: None):
-			instance = AnbimaDataFundsHistoric()
-			instance.logger = mock_logger
-			instance.cls_db = mock_db_session
-			instance.cls_dir_files_management = mock_dir_files_management
-			instance.cls_dates_current = mock_dates_current
-			instance.cls_create_log = mock_create_log
-			instance.cls_dates_br = mock_dates_br
-			instance.date_ref = date(2023, 12, 29)
-			instance.base_url = "https://data.anbima.com.br/fundos"
-			instance.list_fund_codes = ["ABC123", "DEF456"]
-		return instance
-
-	def test_init_with_fund_codes(
-		self, mock_logger: MagicMock, mock_db_session: MagicMock
-	) -> None:
-		"""Test initialization with fund codes.
-
-		Verifies
-		--------
-		- Fund codes list is stored correctly
-		- Default parameters are set
-
-		Parameters
-		----------
-		mock_logger : MagicMock
-			Mock logger instance
-		mock_db_session : MagicMock
-			Mock database session
-
-		Returns
-		-------
-		None
-		"""
-		fund_codes = ["HIST1", "HIST2"]
-		instance = AnbimaDataFundsHistoric(
-			list_fund_codes=fund_codes, logger=mock_logger, cls_db=mock_db_session
-		)
-		assert instance.list_fund_codes == fund_codes
-
-	def test_handle_date_value_replacement(
-		self, funds_historic_instance: AnbimaDataFundsHistoric
-	) -> None:
-		"""Test date value handling for historic data.
-
-		Verifies
-		--------
-		- Invalid date values are replaced with default
-		- Valid dates are preserved
-
-		Parameters
-		----------
-		funds_historic_instance : AnbimaDataFundsHistoric
-			Instance with mocked dependencies
-
-		Returns
-		-------
-		None
-		"""
-		test_cases = [
-			("-", "01/01/2100"),
-			(None, "01/01/2100"),
-			("31/12/2023", "31/12/2023"),
-			("", "01/01/2100"),
-		]
-
-		for input_date, expected in test_cases:
-			result = funds_historic_instance._handle_date_value(input_date)
-			assert result == expected
-
-	def test_transform_data_empty(self, funds_historic_instance: AnbimaDataFundsHistoric) -> None:
-		"""Test data transformation with empty input.
-
-		Verifies
-		--------
-		- Empty list returns empty DataFrame
-		- DataFrame has correct type
-
-		Parameters
-		----------
-		funds_historic_instance : AnbimaDataFundsHistoric
-			Instance with mocked dependencies
-
-		Returns
-		-------
-		None
-		"""
-		result = funds_historic_instance.transform_data([])
-
-		assert isinstance(result, pd.DataFrame)
-		assert len(result) == 0
-
-	def test_transform_data_with_dates(
-		self, funds_historic_instance: AnbimaDataFundsHistoric
-	) -> None:
-		"""Test data transformation with date processing.
-
-		Verifies
-		--------
-		- Date columns are processed through _handle_date_value
-		- Non-date columns remain unchanged
-
-		Parameters
-		----------
-		funds_historic_instance : AnbimaDataFundsHistoric
-			Instance with mocked dependencies
-
-		Returns
-		-------
-		None
-		"""
-		raw_data = [
-			{
-				"FUND_CODE": "ABC123",
-				"DATA_COMPETENCIA": "-",
-				"PL": "R$ 1.000.000,00",
-				"VALOR_COTA": "R$ 100,00",
-			}
-		]
-
-		result = funds_historic_instance.transform_data(raw_data)
-
-		assert isinstance(result, pd.DataFrame)
-		assert len(result) == 1
-		assert result.iloc[0]["DATA_COMPETENCIA"] == "01/01/2100"
-		assert result.iloc[0]["PL"] == "R$ 1.000.000,00"
-
-	def test_parse_raw_file_returns_stringio(
-		self, funds_historic_instance: AnbimaDataFundsHistoric
-	) -> None:
-		"""Test parse_raw_file returns StringIO for compatibility.
-
-		Verifies
-		--------
-		- Method returns StringIO instance
-		- Various response types are accepted
-
-		Parameters
-		----------
-		funds_historic_instance : AnbimaDataFundsHistoric
-			Instance with mocked dependencies
-
-		Returns
-		-------
-		None
-		"""
-		mock_response = MagicMock(spec=SeleniumWebDriver)
-
-		result = funds_historic_instance.parse_raw_file(mock_response)
-
-		assert isinstance(result, StringIO)
-
-	@patch("stpstone.ingestion.countries.br.registries.anbima_data_funds.sync_playwright")
-	def test_get_response_no_fund_codes(
-		self, mock_sync_playwright: MagicMock, funds_historic_instance: AnbimaDataFundsHistoric
-	) -> None:
-		"""Test get_response with empty fund codes.
-
-		Verifies
-		--------
-		- Empty list is returned when no fund codes
-		- Playwright is not invoked
-		- Warning is logged
-
-		Parameters
-		----------
-		mock_sync_playwright : MagicMock
-			Mock sync_playwright context manager
-		funds_historic_instance : AnbimaDataFundsHistoric
-			Instance with mocked dependencies
-
-		Returns
-		-------
-		None
-		"""
-		funds_historic_instance.list_fund_codes = []
-
-		result = funds_historic_instance.get_response()
-
-		assert result == []
-		mock_sync_playwright.assert_not_called()
-
-	def test_run_with_db_insertion(self, funds_historic_instance: AnbimaDataFundsHistoric) -> None:
-		"""Test run method with database insertion.
-
-		Verifies
-		--------
-		- Data is inserted into database when cls_db is provided
-		- All processing steps are executed
-		- No DataFrame is returned
-
-		Parameters
-		----------
-		funds_historic_instance : AnbimaDataFundsHistoric
-			Instance with mocked dependencies
-
-		Returns
-		-------
-		None
-		"""
-		with (
-			patch.object(funds_historic_instance, "get_response") as mock_get_response,
-			patch.object(funds_historic_instance, "transform_data") as mock_transform,
-			patch.object(funds_historic_instance, "standardize_dataframe") as mock_standardize,
-			patch.object(funds_historic_instance, "insert_table_db") as mock_insert,
-		):
-			mock_get_response.return_value = [{"FUND_CODE": "ABC123", "PL": "R$ 1.000.000,00"}]
-			mock_transform.return_value = pd.DataFrame(
-				{"FUND_CODE": ["ABC123"], "PL": ["R$ 1.000.000,00"]}
-			)
-			mock_standardize.return_value = pd.DataFrame(
-				{"FUND_CODE": ["ABC123"], "PL": ["R$ 1.000.000,00"]}
-			)
-
-			result = funds_historic_instance.run()
-
-			mock_get_response.assert_called_once()
-			mock_transform.assert_called_once()
-			mock_standardize.assert_called_once()
-			mock_insert.assert_called_once()
-			assert result is None
-
-	def test_run_without_db_returns_dataframe(
-		self, funds_historic_instance: AnbimaDataFundsHistoric
-	) -> None:
-		"""Test run method returns DataFrame when no database connection.
-
-		Verifies
-		--------
-		- DataFrame is returned when cls_db is None
-		- Database insertion is skipped
-		- All processing steps are executed
-
-		Parameters
-		----------
-		funds_historic_instance : AnbimaDataFundsHistoric
-			Instance with mocked dependencies
-
-		Returns
-		-------
-		None
-		"""
-		funds_historic_instance.cls_db = None
-
-		with (
-			patch.object(funds_historic_instance, "get_response") as mock_get_response,
-			patch.object(funds_historic_instance, "transform_data") as mock_transform,
-			patch.object(funds_historic_instance, "standardize_dataframe") as mock_standardize,
-		):
-			mock_get_response.return_value = [{"FUND_CODE": "ABC123", "PL": "R$ 1.000.000,00"}]
-			expected_df = pd.DataFrame({"FUND_CODE": ["ABC123"], "PL": ["R$ 1.000.000,00"]})
-			mock_transform.return_value = expected_df
-			mock_standardize.return_value = expected_df
-
-			result = funds_historic_instance.run()
-
-			mock_get_response.assert_called_once()
-			mock_transform.assert_called_once()
-			mock_standardize.assert_called_once()
-			assert isinstance(result, pd.DataFrame)
-			assert len(result) == 1
-
-
-# --------------------------
 # Error Handling and Edge Cases
 # --------------------------
 class TestErrorHandling:
-	"""Test error handling and edge cases for all Anbima funds classes."""
+	"""Test error handling and edge cases for AnbimaDataFundsAvailable."""
 
 	def test_funds_available_extraction_exception_handling(self, mock_logger: MagicMock) -> None:
 		"""Test exception handling during fund data extraction.
@@ -1605,145 +832,38 @@ class TestErrorHandling:
 		"""
 		instance = AnbimaDataFundsAvailable(logger=mock_logger)
 
-		# Mock Playwright to simulate extraction failure
 		with patch.object(instance, "_extract_fund_data") as mock_extract:
 			side_effects = [
-				{"NOME_FUNDO": "Fund A"},  # First fund succeeds
-				Exception("Extraction failed"),  # Second fund fails
-				{"NOME_FUNDO": "Fund C"},  # Third fund succeeds
+				{"NOME_FUNDO": "Fund A"},
+				Exception("Extraction failed"),
+				{"NOME_FUNDO": "Fund C"},
 			]
 			mock_extract.side_effect = side_effects
 
-			# Verify the side effect is a list (before it gets converted to iterator)
 			assert isinstance(side_effects, list)
 			assert len(side_effects) == 3
 
-	def test_funds_about_empty_xpath_handling(self, mock_logger: MagicMock) -> None:
-		"""Test handling of empty or missing XPath elements.
-
-		Verifies
-		--------
-		- None is returned for missing elements
-		- Processing continues despite missing data
-		- Logging occurs for missing elements
-
-		Parameters
-		----------
-		mock_logger : MagicMock
-			Mock logger instance
-
-		Returns
-		-------
-		None
-		"""
-		_ = AnbimaDataFundsAbout(logger=mock_logger)
-
-		# Test with mock page that returns no elements
-		mock_page = MagicMock()
-		mock_locator = MagicMock()
-		mock_locator.count.return_value = 0
-		mock_page.locator.return_value = mock_locator
-
-		# Verify the pattern
-		assert mock_locator.count() == 0
-
-	def test_historic_data_missing_columns(self, mock_logger: MagicMock) -> None:
-		"""Test historic data handling with missing columns.
-
-		Verifies
-		--------
-		- Missing columns are handled gracefully
-		- None values are used for missing data
-		- Processing continues despite incomplete data
-
-		Parameters
-		----------
-		mock_logger : MagicMock
-			Mock logger instance
-
-		Returns
-		-------
-		None
-		"""
-		instance = AnbimaDataFundsHistoric(logger=mock_logger)
-
-		# Test transformation with incomplete data
-		raw_data = [
-			{
-				"FUND_CODE": "ABC123",
-				# Missing other columns
-			}
-		]
-
-		result = instance.transform_data(raw_data)
-
-		assert isinstance(result, pd.DataFrame)
-		assert len(result) == 1
-
 	def test_all_classes_inheritance_validation(self) -> None:
-		"""Test that all classes properly inherit from base classes.
+		"""Test that AnbimaDataFundsAvailable properly inherits from base classes.
 
 		Verifies
 		--------
-		- All three classes inherit from ABCIngestionOperations
-		- CoreIngestion and ContentParser are properly initialized
+		- Class inherits from ABCIngestionOperations
 		- Method overrides are correct
 
 		Returns
 		-------
 		None
 		"""
-		# Test Available funds class
 		available = AnbimaDataFundsAvailable()
 		assert isinstance(available, ABCIngestionOperations)
-
-		# Test About funds class
-		about = AnbimaDataFundsAbout()
-		assert isinstance(about, ABCIngestionOperations)
-
-		# Test Historic funds class
-		historic = AnbimaDataFundsHistoric()
-		assert isinstance(historic, ABCIngestionOperations)
-
-	def test_date_handling_edge_cases(self, mock_logger: MagicMock) -> None:
-		"""Test date handling with various edge cases.
-
-		Verifies
-		--------
-		- Various invalid date formats are handled
-		- Default date is used for invalid inputs
-		- Method doesn't raise exceptions
-
-		Parameters
-		----------
-		mock_logger : MagicMock
-			Mock logger instance
-
-		Returns
-		-------
-		None
-		"""
-		instance = AnbimaDataFundsAbout(logger=mock_logger)
-
-		test_cases = [
-			("", "01/01/2100"),
-			(" ", " "),  # Whitespace is not replaced, only "-" and None
-			("invalid", "invalid"),  # Invalid strings that are not "-" are not replaced
-			("2023-13-45", "2023-13-45"),
-			("31/02/2023", "31/02/2023"),  # Invalid date but not "-"
-			("00/00/0000", "00/00/0000"),
-		]
-
-		for test_case, expected in test_cases:
-			result = instance._handle_date_value(test_case)
-			assert result == expected
 
 
 # --------------------------
 # Type Validation Tests
 # --------------------------
 class TestTypeValidation:
-	"""Test type validation for method parameters and return values."""
+	"""Test type validation for AnbimaDataFundsAvailable method parameters and return values."""
 
 	def test_funds_available_method_signatures(self) -> None:
 		"""Test AnbimaDataFundsAvailable method signatures and types.
@@ -1758,7 +878,6 @@ class TestTypeValidation:
 		-------
 		None
 		"""
-		# Check initialization parameters
 		init_annotations = AnbimaDataFundsAvailable.__init__.__annotations__
 		assert "date_ref" in init_annotations
 		assert "logger" in init_annotations
@@ -1766,70 +885,18 @@ class TestTypeValidation:
 		assert "start_page" in init_annotations
 		assert "end_page" in init_annotations
 
-		# Check run method
 		run_annotations = AnbimaDataFundsAvailable.run.__annotations__
 		assert "return" in run_annotations
 
-		# Check get_response method
 		get_response_annotations = AnbimaDataFundsAvailable.get_response.__annotations__
 		assert "return" in get_response_annotations
-
-	def test_funds_about_method_signatures(self) -> None:
-		"""Test AnbimaDataFundsAbout method signatures and types.
-
-		Verifies
-		--------
-		- Method parameters have correct type hints
-		- Return types are correctly annotated
-		- Transformation methods have proper signatures
-
-		Returns
-		-------
-		None
-		"""
-		# Check initialization parameters
-		init_annotations = AnbimaDataFundsAbout.__init__.__annotations__
-		assert "list_fund_codes" in init_annotations
-
-		# Check run method return type
-		run_annotations = AnbimaDataFundsAbout.run.__annotations__
-		assert "return" in run_annotations
-
-		# Check transformation methods
-		char_annotations = AnbimaDataFundsAbout.transform_characteristics_data.__annotations__
-		assert "return" in char_annotations
-
-	def test_funds_historic_method_signatures(self) -> None:
-		"""Test AnbimaDataFundsHistoric method signatures and types.
-
-		Verifies
-		--------
-		- Method parameters have correct type hints
-		- Return types are correctly annotated
-		- All public methods are properly typed
-
-		Returns
-		-------
-		None
-		"""
-		# Check initialization parameters
-		init_annotations = AnbimaDataFundsHistoric.__init__.__annotations__
-		assert "list_fund_codes" in init_annotations
-
-		# Check run method
-		run_annotations = AnbimaDataFundsHistoric.run.__annotations__
-		assert "return" in run_annotations
-
-		# Check data transformation method
-		transform_annotations = AnbimaDataFundsHistoric.transform_data.__annotations__
-		assert "return" in transform_annotations
 
 
 # --------------------------
 # Performance and Mocking Tests
 # --------------------------
 class TestPerformanceAndMocking:
-	"""Test performance optimizations and proper mocking patterns."""
+	"""Test performance optimizations and proper mocking patterns for AnbimaDataFundsAvailable."""
 
 	@pytest.fixture
 	def funds_available_instance_perf(
@@ -1877,8 +944,10 @@ class TestPerformanceAndMocking:
 			instance.end_page = 20
 		return instance
 
-	@patch("stpstone.ingestion.countries.br.registries.anbima_data_funds.time.sleep")
-	@patch("stpstone.ingestion.countries.br.registries.anbima_data_funds.sync_playwright")
+	@patch("stpstone.ingestion.countries.br.registries.anbima_data_funds_available.time.sleep")
+	@patch(
+		"stpstone.ingestion.countries.br.registries.anbima_data_funds_available.sync_playwright"
+	)
 	def test_no_real_network_calls(
 		self,
 		mock_sync_playwright: MagicMock,
@@ -1906,7 +975,6 @@ class TestPerformanceAndMocking:
 		-------
 		None
 		"""
-		# Setup mock Playwright
 		mock_playwright = MagicMock()
 		mock_browser = MagicMock()
 		mock_page = MagicMock()
@@ -1915,7 +983,6 @@ class TestPerformanceAndMocking:
 		mock_playwright.chromium.launch.return_value = mock_browser
 		mock_browser.new_page.return_value = mock_page
 
-		# Mock page interactions
 		mock_locator_total = MagicMock()
 		mock_locator_total.first.inner_text.return_value = "100"
 
@@ -1952,7 +1019,6 @@ class TestPerformanceAndMocking:
 
 			result = funds_available_instance_perf.get_response(timeout_ms=100)
 
-		# Verify mocks were used
 		mock_sync_playwright.assert_called_once()
 		mock_playwright.chromium.launch.assert_called_once()
 		mock_browser.new_page.assert_called_once()
@@ -1979,16 +1045,14 @@ class TestPerformanceAndMocking:
 		"""
 		instance = AnbimaDataFundsAvailable(logger=mock_logger)
 
-		# Test with minimal data
 		small_data = [{"NOME_FUNDO": "Fund A", "LINK_FUNDO": "https://example.com/A"}]
 
 		result = instance.transform_data(small_data)
 
 		assert isinstance(result, pd.DataFrame)
 		assert len(result) == 1
-		# Verify operation is fast by checking it completes quickly
 
-	@patch("stpstone.ingestion.countries.br.registries.anbima_data_funds.pd.DataFrame")
+	@patch("stpstone.ingestion.countries.br.registries.anbima_data_funds_available.pd.DataFrame")
 	def test_mocked_dataframe_operations(
 		self, mock_dataframe: MagicMock, funds_available_instance_perf: AnbimaDataFundsAvailable
 	) -> None:
